Feat: tile size no longer limited to be divisible into canvas size
There was no particular reason for this limitation except it was
easier to implement, but it turns out it wasn't that much easier.
Also, I like the way it looks now, with a partial tile showing that
there's more to the map, and a full tile showing you're at the end.
This is something I've seen recomended in UI design for things like
scrolling lists — show a part of the next item to tell the user that
there's more to see in a particular direction.

* Removed the "windowed" movement that doesn't scroll until you go
  outside a small window around the player location. It's a neat
  effect but I don't think it adds as much in a "cursor" movement
  system.