Refactor: invert input/ui dependency, change input to class
Previously input drove ui, so keypresses would cause a re-render. But
now ui drives input, which will not only be needed now to have input
provide some feedback on the current operation, but also in the future
when I switch things to real time (renders will happen without keypresses).

As part of this, I redesigned the input handler to be a class:

1. Class inheritance allows me to put the common functions (arrow
   handlers, instructions, cursor rendering) into the base class.
2. The class can hold state for the current operation. This was
   possible before but it was hidden and now the state is visible in
   public fields.
3. The class can have additional methods, like rendering.