Feat: real-time simulation loop
Some of this is easier if I just had everything in one big file. It's
harder to think about functionality *and* how to split it up in a way
that I don't get circular dependencies. I should've waited and split
it up after I had a better idea of what structured I needed.

The keyboard input needs to know about the simulation so that it can
start/stop it. The ui needs to know about the simulation so that it
can draw the agents. The simulation needs to know about the ui so that
it can trigger a render on every tick. The keyboard input also needs
to know about the ui so that it can trigger a render on keypress. The
simulation needs to know about the keyboard input so that it can tell
it how to handle keyboard repeat.