Fix: make randInt match python's closed interval; add randRange too
My randInt function returned an integer in the half-open interval [a, b). But
Python's randint function returns in the closed interval [a, b].
Python's randrange is what returns the half-open [a, b), so I renamed
my function to randRange.

But I wanted to use randInt, so I added randInt to return the closed
interval [a, b].